+++ All_lessons/lesson16/lesson16.csproj
@@ -33,8 +33,11 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="PresentationCore" />
+    <Reference Include="PresentationFramework" />
     <Reference Include="System" />
     <Reference Include="System.Core" />
+    <Reference Include="System.Xaml" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
@@ -44,6 +47,7 @@
     <Reference Include="System.Net.Http" />
     <Reference Include="System.Windows.Forms" />
     <Reference Include="System.Xml" />
+    <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="BasicClasses.cs" />
@@ -55,17 +59,23 @@
     <Compile Include="DocumentWindow.Designer.cs">
       <DependentUpon>DocumentWindow.cs</DependentUpon>
     </Compile>
+    <Compile Include="AttributeWindow.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="AttributeWindow.Designer.cs">
+      <DependentUpon>AttributeWindow.cs</DependentUpon>
+    </Compile>
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Form1.Designer.cs">
       <DependentUpon>Form1.cs</DependentUpon>
     </Compile>
-    <Compile Include="AttributeWindow.cs">
-      <SubType>Form</SubType>
+    <Compile Include="GISPanel.cs">
+      <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="AttributeWindow.Designer.cs">
-      <DependentUpon>AttributeWindow.cs</DependentUpon>
+    <Compile Include="GISPanel.Designer.cs">
+      <DependentUpon>GISPanel.cs</DependentUpon>
     </Compile>
     <Compile Include="Layer.cs" />
     <Compile Include="MyFiles.cs" />
@@ -74,11 +84,14 @@
     <EmbeddedResource Include="DocumentWindow.resx">
       <DependentUpon>DocumentWindow.cs</DependentUpon>
     </EmbeddedResource>
+    <EmbeddedResource Include="AttributeWindow.resx">
+      <DependentUpon>AttributeWindow.cs</DependentUpon>
+    </EmbeddedResource>
     <EmbeddedResource Include="Form1.resx">
       <DependentUpon>Form1.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="AttributeWindow.resx">
-      <DependentUpon>AttributeWindow.cs</DependentUpon>
+    <EmbeddedResource Include="GISPanel.resx">
+      <DependentUpon>GISPanel.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
